Updated compiler to check for existing of a module's htdocs directory before attempting to export it

+++ zpt/cdt/compile/Compiler.php
@@ -231,8 +231,10 @@ class Compiler {
         $rmCmd = "rm -r $modTarget";
         exec($rmCmd);
       }
-      $cpCmd = "cp -a $modDocs $target/$modName";
-      exec($cpCmd);
+      if (file_exists($modDocs)) {
+        $cpCmd = "cp -a $modDocs $target/$modName";
+        exec($cpCmd);
+      }
 
       // Compile module models and services
       $modSrc = $module->getPathname() . "/zpt/mod/$modName";
